Move ValidatedJobSubmission into storage
