+++ components/spider-storage/src/cache.rs
@@ -3,6 +3,7 @@ use spider_core::task::TaskIndex;
 pub mod error;
 pub mod io;
 pub mod job;
+pub mod job_submission;
 mod sync;
 pub mod task;
